requirements.txt
# --- Core Web Framework ---
fastapi==0.95.0
uvicorn[standard]==0.27.1
hypercorn==0.14.4

# --- Database & Migrations ---
SQLAlchemy~=2.0.27
psycopg2-binary==2.9.9
alembic~=1.15.1
asyncpg==0.29.0

# --- Async & Background Tasks ---
aiofiles==23.1.0
celery==5.2.7
redis==4.5.4

# --- Auth & Security ---
bcrypt==3.2.2
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv~=1.0.1

# --- HTTP & Testing ---
requests==2.32.4
httpx==0.23.3
pytest==7.2.2

# --- Data Validation ---
pydantic[email]==1.10.14
